# [chore] [exporter/splunk_hec] Remove redundant bufState struct