--- a/exporter/splunkhecexporter/hec_worker_test.go
+++ b/exporter/splunkhecexporter/hec_worker_test.go
@@ -14,7 +14,7 @@ type mockHecWorker struct {
 	failSend bool
 }
 
-func (m *mockHecWorker) send(_ context.Context, _ *bufferState, _ map[string]string) error {
+func (m *mockHecWorker) send(_ context.Context, _ buffer, _ map[string]string) error {
 	if m.failSend {
 		return errHecSendFailed
 	}